initial reorganization of Cell and Node operators
