+++ OperatorCell/Operator.H
@@ -1,0 +1,180 @@
+#ifndef OPERATORCELL_H
+#define OPERATORCELL_H
+
+#include <AMReX_MLCellLinOp.H>
+
+#include "BC/BC.H"
+
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+	// Defined in AMReX
+	void amrex_mg_average (amrex_real* crse,       AMREX_ARLIM_P(crse_lo), AMREX_ARLIM_P(crse_hi),
+			       const amrex_real* fine, AMREX_ARLIM_P(fine_lo), AMREX_ARLIM_P(fine_hi),
+			       const int *tlo, const int *thi,
+			       const int *nc);
+
+	// Defined in AMReX
+	void amrex_mg_interp (amrex_real* fine,       AMREX_ARLIM_P(fine_lo), AMREX_ARLIM_P(fine_hi),
+			      const amrex_real* crse, AMREX_ARLIM_P(crse_lo), AMREX_ARLIM_P(crse_hi),
+			      const int *tlo, const int *thi,
+			      const int *nc);
+#ifdef __cplusplus
+}
+#endif
+
+
+namespace OperatorCell {
+
+class Operator
+	: public amrex::MLCellLinOp
+{
+public:
+
+	friend class MLMG;
+	friend class MLCGSolver;
+
+	Operator ();
+	virtual ~Operator ();
+
+	Operator (const Operator&) = delete;
+	Operator (Operator&&) = delete;
+	Operator& operator= (const Operator&) = delete;
+	Operator& operator= (Operator&&) = delete;
+
+	void define (amrex::Vector<amrex::Geometry> a_geom,
+		     const amrex::Vector<amrex::BoxArray>& a_grids,
+		     const amrex::Vector<amrex::DistributionMapping>& a_dmap,
+		     BC::BC& a_bc,
+		     const amrex::LPInfo& a_info = amrex::LPInfo(),
+		     const amrex::Vector<amrex::FabFactory<amrex::FArrayBox> const*>& a_factory = {});
+
+	// virtual void setLevelBC (int amrlev, const amrex::MultiFab* levelbcdata) override final;
+
+protected:
+	
+	BC::BC *m_bc;
+
+	amrex::Vector<std::unique_ptr<amrex::MLMGBndry> >   m_bndry_sol;
+	amrex::Vector<std::unique_ptr<amrex::BndryRegister> > m_crse_sol_br;
+
+	amrex::Vector<std::unique_ptr<amrex::MLMGBndry> > m_bndry_cor;
+	amrex::Vector<std::unique_ptr<amrex::BndryRegister> > m_crse_cor_br;
+
+	// In case of agglomeration, coarse MG grids on amr level 0 are
+	// not simply coarsened from fine MG grids.  So we need to build
+	// bcond and bcloc for each MG level.
+	using RealTuple = std::array<amrex::Real,2*BL_SPACEDIM>;
+	using BCTuple   = std::array<amrex::BoundCond,2*BL_SPACEDIM>;
+	class BndryCondLoc
+	{
+	public:
+		BndryCondLoc (const amrex::BoxArray& ba, const amrex::DistributionMapping& dm);
+		void setLOBndryConds (const amrex::Geometry& geom, const amrex::Real* dx,
+				      const std::array<BCType,AMREX_SPACEDIM>& lobc,
+				      const std::array<BCType,AMREX_SPACEDIM>& hibc,
+				      int ratio, const amrex::RealVect& a_loc);
+		const BCTuple& bndryConds (const amrex::MFIter& mfi) const {
+			return bcond[mfi];
+		}
+		const RealTuple& bndryLocs (const amrex::MFIter& mfi) const {
+			return bcloc[mfi];
+		}
+	private:
+		amrex::LayoutData<BCTuple>   bcond;
+		amrex::LayoutData<RealTuple> bcloc;
+	};
+	amrex::Vector<amrex::Vector<std::unique_ptr<BndryCondLoc> > > m_bcondloc;
+
+	// used to save interpolation coefficients of the first interior cells
+	mutable amrex::Vector<amrex::Vector<amrex::BndryRegister> > m_undrrelxr;
+
+	// boundary cell flags for covered, not_covered, outside_domain
+	amrex::Vector<amrex::Vector<std::array<amrex::MultiMask,2*AMREX_SPACEDIM> > > m_maskvals;
+
+
+
+	//
+	// functions
+	//
+
+	void updateSolBC (int amrlev, const amrex::MultiFab& crse_bcdata) const;
+	void updateCorBC (int amrlev, const amrex::MultiFab& crse_bcdata) const;
+
+	virtual void applyBC (int amrlev, int mglev, amrex::MultiFab& in, BCMode bc_mode,
+			      const amrex::MLMGBndry* bndry=nullptr, bool skip_fillboundary=false) const;
+
+	virtual void restriction (int, int,
+				  amrex::MultiFab& crse,
+				  amrex::MultiFab& fine) const override final;
+
+	virtual void interpolation (int amrlev, int fmglev,
+				    amrex::MultiFab& fine,
+				    const amrex::MultiFab& crse) const override final;
+
+	virtual void averageDownSolutionRHS (int camrlev,
+					     amrex::MultiFab& crse_sol,
+					     amrex::MultiFab& crse_rhs,
+					     const amrex::MultiFab& fine_sol,
+					     const amrex::MultiFab& fine_rhs) override final;
+
+	virtual void apply (int amrlev, int mglev,
+			    amrex::MultiFab& out,
+			    amrex::MultiFab& in,
+			    BCMode bc_mode,
+			    StateMode s_mode,
+			    const amrex::MLMGBndry* bndry=nullptr) const override final;
+
+
+	virtual void compGrad (int amrlev,
+			       const std::array<amrex::MultiFab*,AMREX_SPACEDIM>& grad,
+			       amrex::MultiFab& sol, amrex::MLLinOp::Location loc) const override final;
+
+	virtual void prepareForSolve () override;
+
+
+	// PURE VIRTUAL METHODS
+
+	virtual void Fapply (int amrlev, int mglev, amrex::MultiFab& out, const amrex::MultiFab& in) const = 0;
+	virtual void Fsmooth (int amrlev, int mglev, amrex::MultiFab& sol, const amrex::MultiFab& rsh, int redblack) const = 0;
+
+	void RegisterNewFab(amrex::Vector<amrex::MultiFab> &input);
+	void RegisterNewFab(amrex::Vector<std::unique_ptr<amrex::MultiFab> > &input);
+	const amrex::FArrayBox & GetFab(const int num, const int amrlev, const int mglev, const amrex::MFIter &mfi) ;
+
+	virtual bool isSingular (int /*amrlev*/) const final { return false; }
+	virtual bool isBottomSingular () const final { return false; }
+
+	virtual amrex::Real getAScalar () const final { return 0.0; }
+	virtual amrex::Real getBScalar () const final { return 0.0; }
+
+	virtual amrex::MultiFab const* getACoeffs (int /*amrlev*/, int /*mglev*/) const final { return nullptr;}
+	virtual std::array<amrex::MultiFab const*,AMREX_SPACEDIM> getBCoeffs (int /*amrlev*/, int /*mglev*/) const final {
+		std::array<amrex::MultiFab const*,AMREX_SPACEDIM> ret;
+		AMREX_D_TERM(ret[0] = nullptr;, ret[1] = nullptr;,ret[2] = nullptr;);
+		return ret;}
+
+	virtual std::unique_ptr<amrex::MLLinOp> makeNLinOp (int /*grid_size*/) const final {
+		Util::Abort("MLABecLaplacian::makeNLinOp: Not implmented");
+		return std::unique_ptr<MLLinOp>{};
+	}
+
+	void averageDownCoeffs ();
+	void averageDownCoeffsSameAmrLevel (amrex::Vector<amrex::MultiFab>&);
+	const amrex::FArrayBox & GetFab(const int num, const int amrlev, const int mglev, const amrex::MFIter &mfi) const;
+
+private:
+
+	void defineAuxData ();
+	void defineBC ();
+
+	int m_num_a_fabs = 0;
+	// m_a_coeffs[index of fab][amrlevel][mglevel]
+	amrex::Vector<amrex::Vector<amrex::Vector<amrex::MultiFab> > > m_a_coeffs; 
+
+};
+}
+
+#endif
